4-dpr/requirements.txt
wget
pytorch
faiss-cpu
transformers